removed unnecessary component, changed Rate component
