--- src/modules/dashboard/components/ExpenditurePage/Streaming/FieldError/FieldError.css
@@ -1,7 +1,0 @@
-.wrapper {
-  composes: main stateError from '~core/Fields/InputStatus/InputStatus.css';
-}
-
-.hidden {
-  display: none;
-}
