--- src/modules/dashboard/components/ExpenditurePage/Streaming/FieldError/FieldError.css.d.ts
@@ -1,2 +1,0 @@
-export const wrapper: string;
-export const hidden: string;
