--- src/modules/dashboard/components/ExpenditurePage/Streaming/FieldError/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FieldError';
