--- src/modules/dashboard/components/ExpenditurePage/Streaming/Rate/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Rate';
