+++ src/modules/dashboard/components/ExpenditurePage/Streaming/RateItem/RateItem.css
