+++ src/modules/dashboard/components/ExpenditurePage/Streaming/RateItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RateItem';
